terminado refactor de todos los modals
