+++ client/src/lib/components/modals/Modal.svelte
